refactor: move submit action from Home to Conversation controller
